Moved logic from app.component to app.service
