+++ app/src/app/app-shell/index.ts
@@ -1,0 +1,1 @@
+export * from './app-shell.module';
