+++ app/src/app/app.component.html
@@ -1,2 +1,2 @@
-<h1 class="sr-only" title="{{pageHeading | translate}}">{{ pageHeading | translate }}</h1>
+<h1 class="sr-only" title="{{pageHeading | async | translate}}">{{ pageHeading | async | translate }}</h1>
 <router-outlet></router-outlet>
