+++ app/src/app/content-plugin/content-services-extension.module.ts
@@ -142,9 +142,9 @@ registerLocaleData(localeSv);
     environment.e2e ? NoopAnimationsModule : BrowserAnimationsModule,
     FormsModule,
     ReactiveFormsModule,
+    ContentModule.forRoot(),
     RouterModule.forChild(CONTENT_ROUTES),
     CoreModule.forChild(),
-    ContentModule.forChild(),
     ExtensionsModule.forChild(),
     CoreExtensionsModule.forChild(),
     SharedModule,
